requirements.txt
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.10.0
cryptography==45.0.5
distlib==0.4.0
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.3.0
google-auth==2.40.3
gunicorn==23.0.0
identify==2.6.12
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.5
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.32.0
whitenoise==6.9.0
